+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Z.ai (GLM) as an OpenAI-compatible provider (#3)
+
 ## [0.5.0] - 2026-04-23
 
 ### Added